backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.3.0"
description = ""
authors = ["Marco Carmona <marcocarmonadev@outlook.com>"]
readme = "README.md"
packages = [{ "include" = "source" }]

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = { extras = ["standard"], version = "^0.30.1" }
pydantic-settings = "^2.3.3"
orjson = "^3.10.5"
python-dateutil = "^2.9.0.post0"
alembic = "^1.13.1"
sqlalchemy-libsql = "^0.1.0"
fastapi = "^0.114.2"

[tool.poetry.scripts]
my-entries-backend = "source.frameworks_and_drivers.web.__main__:main"

[tool.poetry.group.dev.dependencies]
ipython = "^8.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
